+++ src/Sentry.EntityFramework/SentryCommandInterceptor.cs
@@ -54,6 +54,11 @@ namespace Sentry.EntityFramework
         /// </summary>
         public virtual void Log<T>(DbCommand command, DbCommandInterceptionContext<T> interceptionContext)
         {
+            if (string.IsNullOrEmpty(command.CommandText))
+            {
+                return;
+            }
+
             if (interceptionContext.Exception != null)
             {
                 _queryLogger.Log(command.CommandText, BreadcrumbLevel.Error);